+++ packages/plugins/auto-doc/src/ui/components/template-builder/template-builder.svelte
@@ -8,6 +8,7 @@ import ImageElement from '@/ui/components/elements/image-element/image-element.s
 import SignalListElement from '@/ui/components/elements/signal-list-element/signal-list-element.svelte'
 import TableElement from '@/ui/components/elements/table-element/table-element.svelte'
 import TextElement from '@/ui/components/elements/text-element/text-element.svelte'
+import NetworkElement from '@/ui/components/elements/network-element/network-element.svelte'
 import CommunicationElement from '../elements/communication-element/communication-element.svelte'
 
 import { MOVE_BLOCK_DIRECTION } from '@/constants'
@@ -50,6 +51,7 @@ const componentMap: ElementMap = {
 	image: ImageElement,
 	signalList: SignalListElement,
 	table: TableElement,
+	network: NetworkElement,
 	communication: CommunicationElement
 }
 
@@ -160,6 +162,7 @@ function handleContentChange(elementId: string, newContent: string) {
                     <Button variant="outlined" onclick={()=>{addElement("image")}}>Image</Button>
                     <Button variant="outlined" onclick={()=>{addElement("signalList")}}>Signal List</Button>
                     <Button variant="outlined" onclick={() => isCreateTableDialogOpen = true}>Table</Button>
+					<Button variant="outlined" onclick={()=>{addElement("network")}}>Network</Button>
 					<Button variant="outlined" onclick={()=>{addElement("communication")}}>Communication</Button>
                 {/if}
             </div>