+++ packages/plugins/network-explorer/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to the communication explorer plugin will be documented here
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Exports the Network-Explorer Component from the Network Explorer
+### Changed
+- Switched to LegacyTheme
+
 ## [0.0.14] - 2025-11-14
 ### Fixed
 - Set relative path for .css files